requirements.txt
streamlit>=1.44.0
pandas>=2.2.0
python-dotenv>=1.0.1
langchain>=0.3.0
langchain-groq>=0.2.0
pydantic>=2.7.0
PyPDF2>=3.0.1
python-docx>=1.1.2
fastapi>=0.115.0
uvicorn>=0.34.0
python-multipart>=0.0.9
requests>=2.32.0
beautifulsoup4>=4.12.0
